requirements.txt
asgiref             3.9.1
Django              5.2.5
django-cors-headers 4.7.0
djangorestframework 3.16.1
djangorestframework_simplejwt 5.5.1
pillow              11.3.0
pip                 24.3.1
psycopg2-binary     2.9.10
sqlparse            0.5.3
tzdata              2025.2